fix(nmp-kinds): single-source is_replaceable + kind-const consolidation (#1493)
P3 + point-1 of the #1493 architecture-exception audit.

## is_replaceable correctness hazard (CRITICAL)
`nmp_core::kinds::is_replaceable` returned `true` for all 0..=19999 —
treating kind:1/6/7 (note/repost/reaction) as replaceable, the OPPOSITE of
`nostr::Kind::is_replaceable` (used by nmp-store / nmp-nostr-lmdb). Two
reachable functions, same name, opposite answers for a note. Likewise
`is_parameterized_replaceable` wrongly classified the ephemeral
20000..=29999 range as addressable.

Both predicates now live ONCE in the zero-dep Layer-0 `nmp-kinds` crate with
the canonical NIP-01 semantics, bit-identical to `nostr::Kind`:
  is_replaceable               = {0, 3, 41} ∪ 10000..20000
  is_parameterized_replaceable = 30000..40000  (== nostr is_addressable)
The buggy nmp-core copies are deleted; all call sites resolve via
`pub use nmp_kinds::*`. Codex confirmed the ranges and the design (keep
nmp-kinds zero-dep; encode the predicate directly rather than depend on
`nostr`). Hand-coded so it cannot drift; tests assert the kind:1/6/7 and
ephemeral cases that were previously wrong.

## kind-constant fragmentation (D4)
- nmp-wot: import KIND_MUTE_LIST from the registry; drop the stale
  "not yet in the registry" comment and the literal re-declaration.
- nmp-nip01: re-export KIND_SHORT_TEXT_NOTE from the registry; rename the
  crate-internal KIND_SHORT_NOTE callers (no compat alias).
- nmp-marmot: KeyPackage (30443 + legacy 443), group-message (445) and
  welcome (444) kinds were re-declared in five places with a u16/u32 split
  (interest.rs u32, service.rs u16, projection/state.rs u32, tap.rs, ops.rs).
  Collapsed to one canonical u32 set in nmp-kinds; `Kind::Custom`/`as_u16`
  sites cast at the call.
- nmp-nip60: NIP-60/61/88 kinds moved from local u16 consts to the registry
  (u32); the EventBuilder sites cast `as u16` where nostr::Kind requires it.

`is_discovery_kind` (nmp-router) was cited as duplicated but is defined once
and only re-exported — no change. `KIND_LONG_FORM_ARTICLE` / nmp-content are
owned by another lane and left untouched.

+++ crates/nmp-nip60/src/history_event.rs
@@ -89,12 +89,12 @@ pub fn build_history_event(record: &HistoryRecord, keys: &Keys) -> Result<EventB
         })?);
     }
 
-    Ok(EventBuilder::new(Kind::from(KIND_HISTORY), content).tags(tags))
+    Ok(EventBuilder::new(Kind::from(KIND_HISTORY as u16), content).tags(tags))
 }
 
 /// Extract kind:7376 plain `e` tags marked as redeemed nutzap receipts.
 pub fn redeemed_nutzap_ids(event: &Event) -> Vec<EventId> {
-    if event.kind != Kind::from(KIND_HISTORY) {
+    if event.kind != Kind::from(KIND_HISTORY as u16) {
         return Vec::new();
     }
