+++ crates/nmp-nip60/src/token_event.rs
@@ -54,7 +54,7 @@ pub fn build_token_event(record: &TokenRecord, keys: &Keys) -> Result<EventBuild
         nip44::encrypt(keys.secret_key(), &keys.public_key(), json, nip44::Version::V2)
             .map_err(|e| Nip60Error::Nip44(format!("{e}")))?;
 
-    Ok(EventBuilder::new(Kind::from(KIND_TOKEN), content))
+    Ok(EventBuilder::new(Kind::from(KIND_TOKEN as u16), content))
 }
 
 // ─── Decode ────────────────────────────────────────────────────────────────
